fastapi_service/requirements.txt
# FastAPI Service Requirements for Naikoria Tech Academy
# AI Agents & Real-time Features

# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database & ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
databases==0.8.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# AI & LangGraph - Compatible versions
langchain==0.1.7
langchain-openai==0.0.8
langchain-community==0.0.20
langsmith>=0.0.83,<0.1.0
langgraph==0.0.40
openai>=1.10.0,<2.0.0

# Background Tasks & Caching
redis==5.0.1
celery==5.3.4

# Real-time & WebSocket
python-socketio==5.10.0

# HTTP Client for Django communication
httpx==0.25.2
aiohttp==3.9.1

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Monitoring & Logging
structlog==23.2.0